// src/scraper/types/application/Match.ts
export interface MatchParsed {
  league: string
  round: number
  season: string
  matches: Match[]
}

export enum League {
  PremierLeague = 'Premier League'
}

export interface Match {
  teams: string[]
  goals: Array<Goal[]>
  details: Details
  matchFacts: MatchFacts
  matchCards: MatchCard[]
  teamsMatchStats: TeamsMatchStat[]
  playerMatchStats?: PlayerMatchStat[]
}

export interface Details {
  stadium: Stadium
  fullDate: Date
  attendance?: number
  wasGameFinished: boolean
  wasGameCancelled: boolean
  firstHalfStarted: string
  secondHalfStarted: string
  firstHalfEnded: string
  secondHalfEnded: string
  highlights?: string
  referee?: string
  league: League
  round: string
}

export interface Stadium {
  name: string
  capacity: number
  surface: Surface
}

export enum Surface {
  Grass = 'grass'
}

export interface Goal {
  name: string
  ownGoal: boolean | null
  minute: number
  addedTime: number | null
  penalty: boolean
  assistBy?: string
  scoredFor: 0 | 1
}

export interface MatchCard {
  name: string
  time: number
  added: number | null
  card: Card
  description: Description | null
}

export enum Card {
  Red = 'Red',
  Yellow = 'Yellow',
  YellowRed = 'YellowRed'
}

export interface Description {
  localizedKey: string
  defaultText: string
}

export interface MatchFacts {
  manOfTheMatch?: string
  lineups: Lineup[]
}

export interface Lineup {
  formation: string
  starters: string[]
  bench: string[]
  unavailable: Unavailable[]
}

export interface Unavailable {
  name: string
  reason: Reason
  expectedReturn: string
}

export enum Reason {
  Injury = 'injury'
}

export interface PlayerMatchStat {
  player: string
  stats: PlayerMatchStatStat[]
}
// ...
export interface TeamsMatchStat {
  period: Period
  stats: TeamsMatchStatStat[]
}